requirements.txt
amrlib
penman
rdflib
requests
SPARQLWrapper
google-genai
